requirements.txt
# Backend
fastapi==0.109.2
uvicorn[standard]==0.27.1
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
alembic==1.13.1
pydantic>=2.7.4
pydantic-settings>=2.1.0

# LangChain + LangGraph
langchain==0.1.9
langchain-openai==0.0.6
langchain-community==0.0.24
langgraph==0.0.26

# GitHub & CV
PyGithub==2.1.1
httpx==0.26.0
pdfplumber==0.10.4
python-docx==1.1.0
PyPDF2==3.0.1

# Utils
python-multipart==0.0.9
aiofiles==23.2.1
tenacity==8.2.3
structlog==24.1.0

# Frontend
streamlit==1.31.0
requests==2.31.0

# Dev / optional
python-dotenv==1.0.1
